src/test/java/com/aws_s3_wrapper/aws_s3_wrapper/FileUploadServiceImplTest.java
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.mockito.ArgumentMatchers.any;

import com.aws_s3_wrapper.aws_s3_wrapper.S3FileUploadService;
import com.aws_s3_wrapper.grpc.file_upload.FileChunk;
import com.aws_s3_wrapper.grpc.file_upload.FileUploadServiceImpl;
import com.aws_s3_wrapper.grpc.file_upload.UploadStatus;
import com.google.protobuf.ByteString;
import io.grpc.stub.StreamObserver;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

public class FileUploadServiceImplTest {

    @Mock
    private S3FileUploadService s3FileUploadService;

    @Mock
    private StreamObserver<UploadStatus> responseObserver;

    private FileUploadServiceImpl service;

    @BeforeEach
    public void setup() {
        // Initialize mocks created with @Mock annotations
        MockitoAnnotations.initMocks(this);
        // Initialize the service with the mocked S3FileUploadService
        service = new FileUploadServiceImpl(s3FileUploadService);
    }

    @Test
    public void uploadFileTest() {
        // Assuming you have a way to mock or construct a FileChunk instance
        FileChunk mockChunk = FileChunk.newBuilder()
                .setContent(ByteString.copyFromUtf8("test content"))
                .build();

        // Call the method under test
        StreamObserver<FileChunk> requestObserver = service.uploadFile(responseObserver);
        requestObserver.onNext(mockChunk);
        requestObserver.onCompleted();

        // Verify interactions or state changes
        verify(responseObserver, times(1)).onNext(any(UploadStatus.class));
        verify(responseObserver, times(1)).onCompleted();
        // Additional assertions as needed
    }
}
